bug fixed in cliping algorithm

+++ fuzzy_logic/fuzzy_lib/include/rule.h
@@ -12,10 +12,10 @@
 #include<stdarg.h>
 
 void init_rules(rules *rule);
-float MAX(float a, float b);
-float MIN(float a, float b);
-float apply_condition(char* operator, liste liste_values);
-float get_value_in_fuzzy_liste(fuzzy_controler condition, liste fuzzy_result_liste);
+double MAX(double a, double b);
+double MIN(double a, double b);
+double apply_condition(char* operator, liste liste_values);
+double get_value_in_fuzzy_liste(fuzzy_controler condition, liste fuzzy_result_liste);
 liste apply_rules(liste rules_liste, liste fuzzy_result_liste);
 liste creat_rules(fuzzy_controler* rules_tab, int nb_rules);
 void clearRule(rules* r1);
